+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 migration_tools 7.x-2.x  ** - ** - ****
 -----------------------------------------------
+* Adjust params on CheckFor isInPath isSkipFile to make them consistent in
+  param order of needle, haystack.
 * Add constructor to Url.php to create pathing object and some refactoring to
   use and support the new pathing object.
 * Add Migration\HtmlFileBase class.